requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cu111
setuptools==59.5.0
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.3.56
pandas==1.3.3
PyYAML==6.0
scipy==1.7.1
Shapely==1.8.0
tensorboard